move handler creation logic to seperate class
